+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Version info to `/start` command output
+
 ## [0.9.2] - 2023-03-20
 
 ### Fixed